feat: provide static post data
