--- src/app/login/page.tsx
@@ -1,69 +1,0 @@
-'use client';
-
-import { useActionState, useEffect } from 'react';
-import { useFormStatus } from 'react-dom';
-import { login } from '@/app/actions';
-import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { LogIn } from 'lucide-react';
-import { redirect } from 'next/navigation';
-
-
-function SubmitButton() {
-  const { pending } = useFormStatus();
-  return (
-    <Button type="submit" disabled={pending} className="w-full">
-      <LogIn className="mr-2" />
-      {pending ? 'Logging in...' : 'Login'}
-    </Button>
-  );
-}
-
-export default function LoginPage() {
-  const [state, formAction] = useActionState(login, { success: false, errors: {} });
-
-  useEffect(() => {
-    if (state.success) {
-      redirect('/');
-    }
-  }, [state.success]);
-
-  return (
-    <div className="flex items-center justify-center min-h-[70vh]">
-      <div className="w-full max-w-sm">
-        <form action={formAction}>
-          <Card>
-            <CardHeader className="text-center">
-              <CardTitle className="text-2xl">Admin Login</CardTitle>
-              <CardDescription>Enter your password to manage the blog.</CardDescription>
-            </CardHeader>
-            <CardContent className="space-y-4">
-              <div className="space-y-2">
-                <Label htmlFor="password">Password</Label>
-                <Input
-                  id="password"
-                  name="password"
-                  type="password"
-                  required
-                  placeholder="********"
-                  aria-invalid={!!state.errors?.password}
-                  aria-describedby="password-error"
-                />
-                 {state.errors?.password && (
-                  <p id="password-error" className="text-sm font-medium text-destructive">
-                    {state.errors.password[0]}
-                  </p>
-                )}
-              </div>
-            </CardContent>
-            <CardFooter>
-              <SubmitButton />
-            </CardFooter>
-          </Card>
-        </form>
-      </div>
-    </div>
-  );
-}
