+++ src/app/page.tsx
@@ -1,10 +1,8 @@
 import Link from 'next/link';
-import { getPosts, isAuthenticated } from '@/app/actions';
-import { Button } from '@/components/ui/button';
+import { getPosts } from '@/lib/posts';
 import { Card, CardHeader, CardTitle, CardDescription, CardContent } from '@/components/ui/card';
-import { PlusCircle, FileText } from 'lucide-react';
+import { FileText } from 'lucide-react';
 import { Post } from '@/lib/types';
-import { unstable_noStore as noStore } from 'next/cache';
 
 function PostCard({ post }: { post: Post }) {
   return (
@@ -32,36 +30,18 @@ function PostCard({ post }: { post: Post }) {
 }
 
 export default async function Home() {
-  noStore();
   const posts = await getPosts();
-  const authed = await isAuthenticated();
 
   return (
     <div className="space-y-8 animate-fade-in">
       <div className="flex justify-between items-center gap-4">
         <h1 className="text-3xl md:text-4xl font-bold font-headline text-primary">Latest Posts</h1>
-        { authed && (
-          <Button asChild>
-            <Link href="/new">
-              <PlusCircle className="mr-2 h-4 w-4" />
-              New Post
-            </Link>
-          </Button>
-        )}
       </div>
 
       {posts.length === 0 ? (
         <div className="text-center py-20 px-4 border-2 border-dashed rounded-lg bg-card">
           <h2 className="text-xl font-medium">No posts yet</h2>
           <p className="text-muted-foreground mt-2 mb-4">Start by creating your first post.</p>
-           { authed && (
-              <Button asChild>
-                <Link href="/new">
-                  <PlusCircle className="mr-2 h-4 w-4" />
-                  Create a Post
-                </Link>
-              </Button>
-           )}
         </div>
       ) : (
         <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
