+++ src/components/header.tsx
@@ -1,35 +1,7 @@
 import Link from 'next/link';
-import { BookText, LogIn, LogOut } from 'lucide-react';
-import { Button } from './ui/button';
-import { isAuthenticated, logout } from '@/app/actions';
-import { unstable_noStore as noStore } from 'next/cache';
+import { BookText } from 'lucide-react';
 
-async function AuthButton() {
-  const authed = await isAuthenticated();
-
-  if (authed) {
-    return (
-      <form action={logout}>
-        <Button variant="outline" type="submit">
-          <LogOut className="mr-2" />
-          Logout
-        </Button>
-      </form>
-    );
-  }
-
-  return (
-    <Button asChild>
-      <Link href="/login">
-        <LogIn className="mr-2" />
-        Login
-      </Link>
-    </Button>
-  );
-}
-
-export async function Header() {
-  noStore();
+export function Header() {
   return (
     <header className="bg-card/80 backdrop-blur-sm border-b sticky top-0 z-50">
       <div className="container mx-auto px-4 sm:px-6 lg:px-8">
@@ -42,7 +14,6 @@ export async function Header() {
               StaesBlog
             </h1>
           </Link>
-          <AuthButton />
         </div>
       </div>
     </header>
